+++ README.md
@@ -1,1 +1,6 @@
 # pantry-recommender
+
+A ML model that recommends Allrecipes and Jamie Oliver recipes, given an list of ingredients as an input.
+
+### Usage
+In `main.py` modify `user_ingredients` to desired ingredients. Then run ```$ python ./main.py```
