+++ requirements.txt
@@ -1,0 +1,4 @@
+nltk
+scikit-learn
+pandas
+skops
